requirements.txt
google-adk>=1.0.0
pandas>=1.3.0
openpyxl>=3.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
pytest>=7.0.0